get rid of dupe for loops. split out leaderboard data formatting functions into own utils file
